+++ src/Exceptions/MissingCallbackDataException.php
@@ -15,10 +15,20 @@ final class MissingCallbackDataException extends LogicException
     /**
      * @param  array<string>  $requiredKeys
      */
-    public static function make(string $gateway, array $requiredKeys, string $missingKey): self
+    public static function missingKey(string $gateway, array $requiredKeys, string $key): self
     {
         return new self(
-            sprintf('To create %s gateway instance from callback, "%s" are required. "%s" is missing.', $gateway, Arr::join($requiredKeys, ', '), $missingKey)
+            sprintf('To create %s gateway instance from callback, "%s" are required. "%s" is missing.', $gateway, Arr::join($requiredKeys, ', '), $key)
+        );
+    }
+
+    /**
+     * @param  array<string>  $requiredKeys
+     */
+    public static function emptyValue(string $gateway, array $requiredKeys, string $key): self
+    {
+        return new self(
+            sprintf('To create %s gateway instance from callback, "%s" are required. "%s" is empty.', $gateway, Arr::join($requiredKeys, ', '), $key)
         );
     }
 }