+++ tests/Feature/PaymentManagerTest.php
@@ -19,7 +19,7 @@ it('it returns a new instance for the same gateway on each call', function (): v
 it('it returns the same instance when using a fake driver', function (): void {
     paymentManager()->extend(
         'test_gateway',
-        fn (): FakeDriver => app()->make(FakeDriver::class, ['gateway' => 'test_gateway'])
+        fn (): FakeDriver => app()->make(FakeDriver::class, ['gateway' => 'zarinpal'])
     );
 
     $paymentOne = paymentManager()->gateway('test_gateway');